* src/engine/QueryCore.c: 	* src/engine/guid.c: 	* src/engine/kvp-util.c: 	  Fixes for 64-bit architectures.  Fixes bug #113231.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@8346 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/engine/guid.c
@@ -147,7 +147,8 @@ init_from_file(const char *filename, size_t max_size)
 
   file_bytes = init_from_stream(fp, max_size);
 
-  PINFO ("guid_init got %u bytes from %s", file_bytes, filename);
+  PINFO ("guid_init got %llu bytes from %s", (unsigned long long int) file_bytes,
+	 filename);
 
   total += file_bytes;
 
@@ -340,11 +341,12 @@ guid_init(void)
   /* time in secs and clock ticks */
   bytes += init_from_time();
 
-  PINFO ("got %u bytes", bytes);
+  PINFO ("got %llu bytes", (unsigned long long int) bytes);
 
   if (bytes < THRESHOLD)
-    PWARN("only got %u bytes.\n"
-              "The identifiers might not be very random.\n", bytes);
+    PWARN("only got %llu bytes.\n"
+              "The identifiers might not be very random.\n",
+	  (unsigned long long int)bytes);
 
   guid_initialized = TRUE;
 }
